+++ src/velocityTracker.ts
